Fix ambiguous trig type in particle offset math
Cast cos/sin results to CGFloat before multiplying by radius to avoid ambiguous numeric type inference in ParticleBurstView.

+++ App/GlassTheme.swift
@@ -187,8 +187,8 @@ struct ParticleBurstView: View {
         let angle = (Double(index) / Double(particleCount)) * 2 * .pi
         let radius: CGFloat = isAnimating ? CGFloat(80 + (index % 3) * 25) : 0
         return CGSize(
-            width: cos(angle) * radius,
-            height: sin(angle) * radius
+            width: CGFloat(cos(angle)) * radius,
+            height: CGFloat(sin(angle)) * radius
         )
     }
 }
